feat: assembling all togather
