+++ main.go
@@ -1,0 +1,31 @@
+package main
+
+import (
+	"os"
+	"strconv"
+
+	"github.com/shaninalex/financial-analyzer/pkg/datasource"
+	"github.com/shaninalex/financial-analyzer/web"
+)
+
+var (
+	DEBUG        = os.Getenv("DEBUG") // "0" or "1"
+	GURU_API_KEY = os.Getenv("GURU_API_KEY")
+	ALPH_API_KEY = os.Getenv("ALPH_API_KEY")
+	RABBITMQ_URL = os.Getenv("RABBITMQ_URL")
+)
+
+func main() {
+
+	port, err := strconv.Atoi(os.Getenv("APP_PORT"))
+	if err != nil {
+		panic(err)
+	}
+
+	// initialize websocket connection
+	web.Websocket(port)
+
+	// initialize datasource
+	go datasource.Init(GURU_API_KEY, ALPH_API_KEY, RABBITMQ_URL)
+
+}
