--- pkg/datasource/Makefile
@@ -1,10 +1,0 @@
-start:
-	docker compose up -d --build
-
-down:
-	docker compose down
-
-rebuild_datasource_service:
-	docker compose up -d --no-deps --build datasource
-
-restart: down start
