--- pkg/datasource/README.md
@@ -1,4 +1,0 @@
-## Datasource
-
-This is the datasource service. It responsible for requests to different 
-Is, store data in Redis, handle incorrect requests, managing user API keys.
