+++ pkg/datasource/datasource.go
@@ -2,7 +2,7 @@
 This package used for getting information from financials APIs such as
 GuruFocus, Alphavantage etc.
 */
-package main
+package datasource
 
 // Usage example: data, err: := app.Datasource.Gurufocus.summary("AAPL")
 type Datasource struct {
