+++ pkg/datasource/datasource_test.go
@@ -1,4 +1,4 @@
-package main
+package datasource
 
 import (
 	"os"
