--- pkg/datasource/docker-compose.yml
@@ -1,44 +1,0 @@
-version: '3.7'
-
-volumes:
-  rabbitmq_data:
-
-services:
-  datasource:
-    build: ./
-    container_name: finapp_datasource
-    restart: on-failure
-    ports:
-      - 8002:8002
-    environment:
-      - APP_PORT=8002
-      - GURU_API_KEY=random_key
-      - ALPH_API_KEY=demo
-      - RABBITMQ_URL=amqp://guest:guest@rabbitmq:5672/
-    depends_on:
-      rabbitmq_post_initialize:
-        condition: service_completed_successfully
-
-  rabbitmq:
-    image: rabbitmq:3-management-alpine
-    container_name: finapp_rabbitmq
-    ports:
-      - 5672:5672
-      - 15672:15672
-    volumes:
-      - rabbitmq_data:/var/lib/rabbitmq/mnesia/
-    healthcheck:
-      test: rabbitmq-diagnostics -q ping
-      interval: 5s
-      timeout: 5s
-      retries: 3
-
-  rabbitmq_post_initialize:
-    build: ../config/rabbitmq
-    container_name: finapp_rabbitmq_post_initialize
-    environment:
-      - RABBITMQ_HOST=rabbitmq
-      - RABBITMQ_PORT=5672
-    depends_on:
-      rabbitmq:
-        condition: service_healthy
