+++ pkg/datasource/middleware_requests.go
@@ -1,1 +1,1 @@
-package main
+package datasource
