+++ pkg/datasource/service_alphavantage.go
@@ -1,4 +1,4 @@
-package main
+package datasource
 
 import (
 	"encoding/json"
