+++ pkg/datasource/service_alphavantage_test.go
@@ -1,4 +1,4 @@
-package main
+package datasource
 
 import (
 	"testing"
