+++ pkg/datasource/types_gurufocus.go
@@ -1,4 +1,4 @@
-package main
+package datasource
 
 // This object includes increadible amount of details and for not I will not
 // define them all. More details can be found in the official documentation:
