--- pkg/notifications/Dockerfile
@@ -1,11 +1,0 @@
-FROM golang:1.21.3
-WORKDIR /build
-COPY . ./
-RUN go mod tidy
-RUN CGO_ENABLED=0 GOOS=linux go build -a -installsuffix cgo -o notification .
-
-FROM alpine:latest  
-RUN apk --no-cache add ca-certificates
-WORKDIR /root/
-COPY --from=0 /build/notification ./
-CMD ["./notification"]
