--- pkg/notifications/utils.go
@@ -1,11 +1,0 @@
-package main
-
-import (
-	"log"
-)
-
-func failOnError(err error) {
-	if err != nil {
-		log.Panicf("%s", err)
-	}
-}
